Refactor the <v-card>-component
