--- resources/frontend/src/app/components/card/body/body.js
@@ -1,24 +1,0 @@
-import ComponentMixin from './../../../mixins/component';
-import ElementMixin from './../../../mixins/element';
-import SlotMixin from './../../../mixins/slot';
-
-export default {
-  /**
-   * The name of the component
-   */
-  name: 'card-body',
-
-  /**
-   * The element name of the component.
-   */
-  element: 'body',
-
-  /**
-   * The mixins used to extend the component.
-   */
-  mixins: [
-    ComponentMixin,
-    ElementMixin,
-    SlotMixin,
-  ],
-};
