--- resources/frontend/src/app/components/card/body/body.styl
@@ -1,7 +1,0 @@
-@import './../../../../assets/stylus/imports.styl'
-
-.card
-  +has('body')
-    display: flex
-    padding: 1rem
-    width: 100%
