--- resources/frontend/src/app/components/card/body/body.vue
@@ -1,3 +1,0 @@
-<style src="./body.styl" lang="stylus" module></style>
-<template src="./body.html"></template>
-<script src="./body.js"></script>
