--- resources/frontend/src/app/components/card/card.html
@@ -1,5 +1,0 @@
-<div :class="classNames">
-  <slot>
-    {{ content }}
-  </slot>
-</div>
