--- resources/frontend/src/app/components/card/card.js
@@ -1,19 +1,0 @@
-import ComponentMixin from './../../mixins/component';
-import BlockMixin from './../../mixins/block';
-import SlotMixin from './../../mixins/slot';
-
-export default {
-  /**
-   * The name of the component
-   */
-  name: 'card',
-
-  /**
-   * The mixins used to extend the component.
-   */
-  mixins: [
-    ComponentMixin,
-    BlockMixin,
-    SlotMixin,
-  ],
-};
