--- resources/frontend/src/app/components/card/card.styl
@@ -1,11 +1,0 @@
-@import './../../../assets/stylus/imports.styl'
-
-.card
-  background-color: palette('White')
-  border-radius: .125rem
-  box-shadow: $box_shadow_minimum
-  display: flex
-  width: 100%
-  +variant('inline')
-    display: inline-flex
-    width: auto
