+++ resources/frontend/src/app/components/card/index.js
@@ -1,0 +1,5 @@
+import * as Card from './card';
+import * as CardBody from './card-body';
+
+export const VCard = Card;
+export const VCardBody = CardBody;
