+++ resources/frontend/src/app/components/drawer/drawer.styl
@@ -6,7 +6,6 @@
   flex-direction: column
   height: 100%
   left: 0
-  //left: -17.5rem
   position: fixed
   top: 0
   transition: left $transition_speed ease-in-out
