+++ resources/frontend/src/app/pages/not-found/index/index.js
@@ -6,8 +6,7 @@
  */
 
 import * as VLayout from './../../../layouts/base/base.vue';
-import * as VCard from './../../../components/card/card.vue';
-import * as VCardBody from './../../../components/card/body/body.vue';
+import { VCard, VCardBody } from './../../../components/card';
 import * as VGrid from './../../../components/grid/grid.vue';
 import * as VRow from './../../../components/row/row.vue';
 import * as VCol from './../../../components/col/col.vue';
