+++ resources/frontend/src/app/pages/scammers/index/index.js
@@ -7,8 +7,7 @@ import * as VCol from './../../../components/col/col.vue';
 import * as VLayout from './../../../layouts/base/base.vue';
 import * as VTable from './../../../components/table/table.vue';
 import * as VPagination from './../../../components/pagination/pagination.vue';
-import * as VCard from './../../../components/card/card.vue';
-import * as VCardBody from './../../../components/card/body/body.vue';
+import { VCard, VCardBody } from './../../../components/card';
 import * as VForm from './../../../components/form/form.vue';
 import * as VTextField from './../../../components/text-field/text-field.vue';
 import * as VInputGroup from './../../../components/input-group/input-group.vue';
